Added the final structure for the user endpoint
