+++ src/routes/users/index.ts
@@ -1,0 +1,13 @@
+
+import { Router, Request, Response, NextFunction } from "express";
+import externalRouter from "./external";
+import internalRouter from "./internal";
+
+const userRouter = Router();
+
+userRouter.use("/internal", internalRouter)
+userRouter.use("/", externalRouter)
+
+
+export default userRouter;
+
