+++ src/utils/helperFunctions.ts
@@ -1,0 +1,14 @@
+import bcrypt from "bcrypt";
+
+export function encrypt(password: string, rounds: number = 10) {
+  return new Promise((resolve, reject) => {
+    bcrypt.hash(password, rounds, (err, hash) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(hash);
+      }
+    });
+  });
+}
+
